+++ integration-tests/functions/gatsby-config.js
@@ -1,0 +1,9 @@
+module.exports = {
+  flags: {
+    FUNCTIONS: true,
+  },
+  siteMetadata: {
+    title: "functions",
+  },
+  plugins: [`gatsby-plugin-gatsby-cloud`],
+}