+++ packages/gatsby-source-graphql/CHANGELOG.md
@@ -3,6 +3,28 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [3.5.0-next.0](https://github.com/gatsbyjs/gatsby/compare/gatsby-source-graphql@3.4.0-next.2...gatsby-source-graphql@3.5.0-next.0) (2021-04-23)
+
+**Note:** Version bump only for package gatsby-source-graphql
+
+# [3.4.0-next.2](https://github.com/gatsbyjs/gatsby/compare/gatsby-source-graphql@3.4.0-next.1...gatsby-source-graphql@3.4.0-next.2) (2021-04-22)
+
+### Bug Fixes
+
+- **deps:** update minor and patch for gatsby-source-graphql ([#28360](https://github.com/gatsbyjs/gatsby/issues/28360)) ([604c25f](https://github.com/gatsbyjs/gatsby/commit/604c25fe7a7c15c7a5ae8fa25b538c1631df330a))
+
+# [3.4.0-next.1](https://github.com/gatsbyjs/gatsby/compare/gatsby-source-graphql@3.4.0-next.0...gatsby-source-graphql@3.4.0-next.1) (2021-04-19)
+
+**Note:** Version bump only for package gatsby-source-graphql
+
+# [3.4.0-next.0](https://github.com/gatsbyjs/gatsby/compare/gatsby-source-graphql@3.3.0-next.1...gatsby-source-graphql@3.4.0-next.0) (2021-04-09)
+
+**Note:** Version bump only for package gatsby-source-graphql
+
+# [3.3.0-next.1](https://github.com/gatsbyjs/gatsby/compare/gatsby-source-graphql@3.3.0-next.0...gatsby-source-graphql@3.3.0-next.1) (2021-04-07)
+
+**Note:** Version bump only for package gatsby-source-graphql
+
 # [3.3.0-next.0](https://github.com/gatsbyjs/gatsby/compare/gatsby-source-graphql@3.2.0-next.0...gatsby-source-graphql@3.3.0-next.0) (2021-03-25)
 
 **Note:** Version bump only for package gatsby-source-graphql